% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{add_border}
\alias{add_border}
\title{Add border to cells for forestplot}
\usage{
add_border(
  plot,
  row = NULL,
  col = NULL,
  part = c("header", "body"),
  where = c("bottom", "left", "top", "right"),
  linewidth = 0.5,
  ...
)
}
\arguments{
\item{plot}{A forest plot object.}

\item{row}{A numeric value or vector indicating row number to add border.
This is corresponding to the data row number. Remember to account for any
text inserted. A border will be drawn to all rows if this is omitted.}

\item{col}{A numeric value or vector indicating the columns to add border.
A border will be drawn to all columns if this is omitted.}

\item{part}{The border will be added to "body" or "header" (default) .}

\item{where}{Where to draw the border of the cell, possible values are
"bottom" (default), "left", "top" and "right"}

\item{linewidth}{line width, default 0.5.}

\item{...}{more arguments.}
}
\value{
A gtable object.
}
\description{
Add border to cells for forestplot
}
