% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aps.R
\docType{data}
\name{aps}
\alias{aps}
\title{APS data}
\format{
A data.frame with 508 rows and 10 variables.
}
\source{
Hosmer, D.W., Lemeshow, S. and Sturdivant, R.X. (2013)
Applied Logistic Regression, 3rd ed., New York: Wiley
}
\usage{
aps
}
\description{
aps dataset.
}
\examples{
head(aps, 10)
}
\keyword{datasets}
