% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\alias{cal.data.frame}
\alias{cal.nmtask}
\title{Draw calibrate curve}
\usage{
cal(data, outcome = NULL, predictors = NULL, newdata = NULL, B = 1000, ...)

\method{cal}{data.frame}(data, outcome = NULL, predictors = NULL, newdata = NULL, B = 1000, ...)

\method{cal}{nmtask}(data, outcome = NULL, predictors = NULL, newdata = NULL, B = 1000, ...)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.}

\item{B}{B is an upper limit on the number of resamples for which information
is printed about which variables were selected in each model re-fit.}

\item{...}{further arguments.}
}
\description{
Draw calibrate curve
}
