% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefs}
\alias{coefs}
\title{Model coefficients of logistic regression}
\usage{
coefs(
  ...,
  multivariable = TRUE,
  effect.values = c("net", "b", "se", "OR", "p"),
  conf.level = 0.95,
  conf.brackets = "()",
  conf.separator = NULL,
  digits.pvalue = 3,
  digits.effect = 2,
  ref.value = "Reference",
  filename = ""
)
}
\arguments{
\item{...}{one or more object of 'nmtask' or 'glm'.}

\item{multivariable}{logical variable indicating whether to perform univariable
logistic analysis or multivariable logistic analysis (default).}

\item{effect.values}{Effect value.}

\item{conf.level}{Effect value level, defalut 0.95.}

\item{conf.brackets}{Confidence interval bracket, default '('.}

\item{conf.separator}{Confidence interval separator.}

\item{digits.pvalue}{digits for p value, default 3.}

\item{digits.effect}{digits for effect value, default 2.}

\item{ref.value}{Reference Category name, default 'Reference'.}

\item{filename}{filename, if you want to save to word.}
}
\value{
a list.
}
\description{
Model coefficients of logistic regression
}
