% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Comparision of nomogram models}
\usage{
compare(..., cutoff = NULL, digits = 3, filename = "")
}
\arguments{
\item{...}{nomogram models.}

\item{cutoff}{cutoff values for risk categories NRI.if is NULL, the function will compute continuous NRI.}

\item{digits}{digits, default 3.}

\item{filename}{filename, if you want to save to word.}
}
\value{
a data frame.
}
\description{
Comparision of nomogram models
}
\details{
The function computes the categorical and continuous net reclassification
improvement (NRI) and integrated discrimination improvement (IDI). A reclassification
table indicates the number of individuals who move to another risk category or
remain in the same risk category as a result of updating the risk model.
Categorical NRI equal to x\% means that compared with individuals without outcome,
individuals with outcome were almost x\% more likely to move up a category than
down. The function also computes continuous NRI, which does not require any
discrete risk categories and relies on the proportions of individuals with
outcome correctly assigned a higher probability and individuals without outcome
correctly assigned a lower probability by an updated model compared with the
initial model. IDI equal to x\% means that the difference in average predicted
risks between the individuals with and without the outcome increased by x\% in
the updated model.
}
