% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.data.frame}
\alias{evaluate.nmtask}
\title{Evaluate model performance}
\usage{
evaluate(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  digits = 3,
  filename = "",
  ...
)

\method{evaluate}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  digits = 3,
  filename = "",
  ...
)

\method{evaluate}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  digits = 3,
  filename = "",
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.
is printed about which variables were selected in each model re-fit.}

\item{digits}{digits, defalut 3.}

\item{filename}{file name to save performance metrics.}

\item{...}{further arguments.}
}
\description{
Evaluate model performance
}
