% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forest}
\alias{forest}
\title{Draw forest plot}
\usage{
forest(
  data,
  estimate = NULL,
  lower = NULL,
  upper = NULL,
  sizes = 0.4,
  ref.line = 1,
  col.width = 2,
  graph.pos = NULL,
  fontfamily = "serif",
  fontsize = 12,
  backgroud = NULL,
  backgroup.odd = "#eff3f2",
  backgroup.even = "white",
  est.color = "#0093bd",
  ci.color = est.color,
  est.shape = 16,
  graph.width = 25,
  axis.ticks = NULL,
  axis.ticks.digits = 1L,
  axis.labels = NULL,
  footnote = NULL,
  ...
)
}
\arguments{
\item{data}{Data (a data frame, an object from \link{coefs},\link{nmtask} or \link{glm}) to be displayed in the forest plot}

\item{estimate}{Point estimation.}

\item{lower}{Lower bound of the confidence interval.}

\item{upper}{Upper bound of the confidence interval.}

\item{sizes}{Size of the point estimation box, can be a unit, vector or a list.}

\item{ref.line}{X-axis coordinates of zero line, default is 1.}

\item{col.width}{Adjust the column width with space, default 2 spaces.}

\item{graph.pos}{Column number of the data the CI will be displayed.}

\item{fontfamily}{The font family.}

\item{fontsize}{The size of text.}

\item{backgroud}{backgroud}

\item{backgroup.odd}{Odd row background color, It can be used when backgroud is NULL.}

\item{backgroup.even}{Even row background color, It can be used when backgroud is NULL.}

\item{est.color}{Color fill the point estimation.}

\item{ci.color}{Color of the CI.}

\item{est.shape}{Shape of the point estimation.}

\item{graph.width}{Adjust the graph width with space, default 25 spaces.}

\item{axis.ticks}{Set X-axis tick-marks point.}

\item{axis.ticks.digits}{Number of digits for the x-axis, default is 1L.}

\item{axis.labels}{Labels for the arrows, string vector of length two (left
and right).}

\item{footnote}{Footnote for the forest plot, will be aligned at left bottom
of the plot. Please adjust the line length with line break to avoid the
overlap with the arrow and/or x-axis.}

\item{...}{more arguments pass to \link{forest_theme} function.}
}
\value{
A gtable object.
}
\description{
Draw forest plot
}
\examples{
tk <- nmtask(train.data = HCC,
             outcome    = "status",
             predictors = c("Grade", "AJCC_T:AJCC_M", "Surg_Prim", "metastasis"))
forest(tk)
forest(coefs(tk))
forest(tk, axis.labels = c("Favors nagative", "Favors pasitive"))

forest(tk) |>
  add_border(row = 1, part = "header", where = "bottom")

forest(tk,
       backgroud = "white",
       axis.ticks = c(0, 1, 2, 3, 4, 5)) |>
  add_border(row = 1, part = "header", where = "bottom")

model <- glm(status ~ Grade + AJCC_T + AJCC_M + AJCC_N + Surg_Prim + metastasis,
             data = HCC,
             family = binomial())
forest(model)
forest(coefs(model))
}
