% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HCC.R
\docType{data}
\name{HCC}
\alias{HCC}
\title{Hepatocellular carcinoma (HCC) data}
\format{
A data.frame with 15985 rows and 14 variables.
\itemize{
\item ID: Patient's ID.
\item Race: Race.
\item Sex: Gender.
\item Age: Age at diagnosis (years).
\item Grade: Histological grading.
\item AJCC_T: T stage of AJCC.
\item AJCC_N: N stage of AJCC.
\item AJCC_M: M stage of AJCC.
\item Surg_Prim: Whether the primary lesion should be treated by surgery.
\item tumor_size: Tumor size (mm).
\item lymph_nodes: Number of lymph nodes.
\item metastasis: Metastasis.
\item time: Survival time (month).
\item status: Censoring status.
}
}
\usage{
HCC
}
\description{
Survival in patients with HCC from the the Surveillance, Epidemiology, and End Results (SEER) dataset.
}
\examples{
head(HCC, 10)
}
\keyword{datasets}
