% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoslem_test.R
\name{hoslem_test}
\alias{hoslem_test}
\alias{hoslem_test.data.frame}
\alias{hoslem_test.nmtask}
\title{Hosmer-Lemeshow Test}
\usage{
hoslem_test(data, outcome = NULL, predictors = NULL, g = 10)

\method{hoslem_test}{data.frame}(data, outcome = NULL, predictors = NULL, g = 10)

\method{hoslem_test}{nmtask}(data, outcome = NULL, predictors = NULL, g = 10)
}
\arguments{
\item{data}{data.}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{g}{number of bins to use to calculate quantiles.}
}
\description{
Hosmer-Lemeshow Test
}
