% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{insert_text}
\alias{insert_text}
\title{Insert text to forest plot}
\usage{
insert_text(
  plot,
  text,
  row = NULL,
  col = NULL,
  part = c("header", "body"),
  just = c("center", "left", "right"),
  before = TRUE,
  fontfamily = "serif",
  fontsize = 12,
  fontface = "bold",
  padding = grid::unit(1, "mm"),
  ...
)
}
\arguments{
\item{plot}{A forest plot object.}

\item{text}{A character or expression vector.}

\item{row}{Row to insert the text, this will be ignored if the part is "header".}

\item{col}{A numeric value or vector indicating the columns the text will be
added. The text will span over the column if a vector is given.}

\item{part}{Part to insert text, body or header (default).}

\item{just}{The justification of the text, "center" (default), "left" or "right".}

\item{before}{Indicating the text will be inserted before or after the row.}

\item{fontfamily}{font family, default serif.}

\item{fontsize}{font size, default 12.}

\item{fontface}{font face, default bold.}

\item{padding}{padding}

\item{...}{more argumentd.}
}
\value{
A gtable object.
}
\description{
Insert text to forest plot
}
