% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{label_category}
\alias{label_category}
\title{Set labels for category of nomogram}
\usage{
label_category(x, variable, ...)
}
\arguments{
\item{x}{a nomogram object from \link{nom} function}

\item{variable}{variable name from nomogram.}

\item{...}{"category=label"}
}
\value{
a nomogram.
}
\description{
Set labels for category of nomogram
}
\seealso{
\link{label_variable}
}
