% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{label_variable}
\alias{label_variable}
\title{Set labels for variable of nomogram}
\usage{
label_variable(x, ...)
}
\arguments{
\item{x}{a nomogram object from nom() function}

\item{...}{"vairable=label"}
}
\value{
a nomogram.
}
\description{
Set labels for variable of nomogram
}
\seealso{
\link{label_category}
}
