% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{nom}
\alias{nom}
\alias{nom.data.frame}
\alias{nom.nmtask}
\title{Draw a nomogram}
\usage{
nom(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  show.points = FALSE,
  show.model = FALSE,
  ...
)

\method{nom}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  show.points = FALSE,
  show.model = FALSE,
  ...
)

\method{nom}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  show.points = FALSE,
  show.model = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{points.label}{a character string giving the axis label for the points scale.}

\item{total.points.label}{a character string giving the axis label for the total points scale.}

\item{funlabel}{label for fun axis.}

\item{maxscale}{default maximum point score is 100.}

\item{xfrac}{fraction of horizontal plot to set aside for axis titles.}

\item{font.size}{font size.}

\item{fun.at}{function values to label on axis.}

\item{show.points}{show points, default FALSE..}

\item{show.model}{show model, default FALSE.}

\item{...}{settings of variables to use in constructing axes.}
}
\description{
The nomogram can be interpreted as follows: (1) for each  variable,
draw a straight line up to the points axis to determine the points for that
variable, (2) repeat this process for each variable, (3) add the points for
all variables and locate the sum on the total points axis, and (4) draw a
straight line from total points down to risk.
}
