% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf}
\alias{perf}
\title{Evaluate model performance}
\usage{
perf(
  ...,
  newdata = NULL,
  cutoff = "best",
  probability = TRUE,
  digits = 3,
  filename = ""
)
}
\arguments{
\item{...}{one or more object of 'nmtask' or 'glm'.}

\item{newdata}{new data for verification.}

\item{cutoff}{Cutoff is the prediction probability of logistic regression model,
between 0 and 1. By default, the best cut-off value is calculated according to
the maximum Yodon index.}

\item{probability}{Whether to calculate the probability of the best cutoff value.}

\item{digits}{digits, default 3.}

\item{filename}{filename, if you want to save to word.}
}
\description{
Evaluate the performance of the model, including the area under the ROC curve,
accuracy, sensitivity, specificity, positive predictive value, negative
predictive value, and Brier score.
}
\details{
About Brier score:

The Brier score is a proper score function that measures the accuracy of
probabilistic predictions. It is applicable to tasks in which predictions
must assign probabilities to a set of mutually exclusive discrete outcomes.
The set of possible outcomes can be either binary or categorical in nature,
and the probabilities assigned to this set of outcomes must sum to one (where
each individual probability is in the range of 0 to 1).

The lower the Brier score is for a set of predictions, the better the
predictions are calibrated. Note that the Brier score, in its most common
formulation, takes on a value between zero and one, since this is the largest
possible difference between a predicted probability (which must be between zero
and one) and the actual outcome (which can take on values of only 0 and 1).
(In the original (1950) formulation of the Brier score, the range is double,
from zero to two.)
}
\examples{
# Examples see in [nmtask] function.
}
\references{
Brier, G. W. (1950) Verification of forecasts expressed in terms of probability. Monthly Weather Review, 78, 1-3.
}
