% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{print.coefs}
\alias{print.coefs}
\title{Print 'coefs'}
\usage{
\method{print}{coefs}(x, ...)
}
\arguments{
\item{x}{an object of 'coefs'}

\item{...}{more.}
}
\description{
Print 'coefs'
}
\keyword{internal}
