% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassification.R
\name{print.compare}
\alias{print.compare}
\title{Print object}
\usage{
\method{print}{compare}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{more.}
}
\description{
Print object
}
\keyword{internal}
