% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoslem_test.R
\name{print.hoslem}
\alias{print.hoslem}
\title{Print hoslem}
\usage{
\method{print}{hoslem}(x, ...)
}
\arguments{
\item{x}{a object of hoslem}

\item{...}{more.}
}
\description{
Print hoslem
}
\keyword{internal}
