% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-nmplot.R
\name{print.nmplot}
\alias{print.nmplot}
\title{Print plots}
\usage{
\method{print}{nmplot}(x, ...)
}
\arguments{
\item{x}{an object of 'nmplot'.}

\item{...}{more.}
}
\description{
Print plots
}
\keyword{internal}
