% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-nmtask.R
\name{print.nmtask}
\alias{print.nmtask}
\title{Print nmtask}
\usage{
\method{print}{nmtask}(x, ...)
}
\arguments{
\item{x}{an object of class 'nmtask'.}

\item{...}{further arguments.}
}
\description{
Print nmtask
}
