% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{print.nomologit}
\alias{print.nomologit}
\title{Print nomogram}
\usage{
\method{print}{nomologit}(x, ...)
}
\arguments{
\item{x}{an object of 'nomologit'.}

\item{...}{more arguments}
}
\description{
Print nomogram
}
\keyword{internal}
