% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{print.performance}
\alias{print.performance}
\title{Print performance}
\usage{
\method{print}{performance}(x, ...)
}
\arguments{
\item{x}{a object of performance}

\item{...}{more.}
}
\description{
Print performance
}
\keyword{internal}
