% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{roc.data.frame}
\alias{roc.nmtask}
\title{Draw ROC curve}
\usage{
roc(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{roc}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{roc}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.
is printed about which variables were selected in each model re-fit.}

\item{xlab}{defaults to "Predicted x-units Survival" or to a suitable label for other models.}

\item{ylab}{defaults to "Fraction Surviving x-units" or to a suitable label for other models.}

\item{...}{further arguments.}
}
\description{
Draw ROC curve
}
