% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_exec}
\alias{roc_exec}
\title{Build ROCs}
\usage{
roc_exec(
  data,
  outcome,
  exposure,
  positive = NULL,
  threshold = "best",
  combine = FALSE,
  combine.only = FALSE,
  ci = TRUE,
  ci.method = c("delong", "bootstrap"),
  ci.sep = NULL,
  ci.branket = c("(", "["),
  smooth = FALSE,
  smooth.args = list(),
  digits = 2,
  percent = FALSE,
  language = NULL,
  table.number = NULL,
  progress = "win",
  boot.n = 1000,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{the outcome variable name contain in data, which can be a factor
or a numerical variable, but must be a binary variable.}

\item{exposure}{one or more exposure variable names are included in data.}

\item{positive}{in which  positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{threshold}{threshold, default "best", the optimal cut-off is the threshold
that calculating by youden index.}

\item{combine}{a logical, indicating whether to use multiple exposure variables
for joint prediction using logistic regression.}

\item{combine.only}{a logical, indicating whether to show only the results of
joint predictions.}

\item{ci}{a logical, indicating whether to show ci.}

\item{ci.method}{the method to use, either “delong” or “bootstrap”. The
first letter is sufficient. If omitted, the appropriate method is selected as
explained in details.}

\item{ci.sep}{the separator for CI.}

\item{ci.branket}{the branket for CI.}

\item{smooth}{a logical, indicating whether to smooth a ROC curve.}

\item{smooth.args}{further arguments for smooth, see [pROC::smooth] function.}

\item{digits}{digits, default 2.}

\item{percent}{percenttage, default FALSE.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{table.number}{table number.}

\item{progress}{the name of progress bar to display. Typically “none”, “win”,
“tk” or “text”.}

\item{boot.n}{the number of bootstrap replicates, default 1000.}

\item{seed}{seed, default 1234.}

\item{...}{unused arguments.}
}
\value{
a data frame.
}
\description{
Build ROCs
}
\keyword{internal}
