% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{set_variable_labels}
\alias{set_variable_labels}
\title{Set labels for nomogram}
\usage{
set_variable_labels(x, ...)
}
\arguments{
\item{x}{a nomogram object from nom() function}

\item{...}{"vairable=label"}
}
\value{
a nomogram.
}
\description{
Set labels for nomogram
}
\examples{
nm <- nom(data = aps,
          outcome = "elope",
          predictors = c("age", "gender", "place3"))

nm |>
  set_variable_labels(age = "Age (years)") |>
  set_variable_labels(gender = "Gender", place3 = "Placement") |>
  set_category_labels("Gender", Male = "M", Female = "F") |>
  set_category_labels("Placement", OutDay = "Outday")
}
\seealso{
\link{set_category_labels}
}
