% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP.R
\name{shap_dependence}
\alias{shap_dependence}
\title{SHAP dependence plots}
\usage{
shap_dependence(x, colors = "#3b528b", ...)
}
\arguments{
\item{x}{An object from \link{shap} function.}

\item{colors}{Color used to fill the bars (only used if bars are shown).}

\item{...}{More arguments.}
}
\value{
An object of class ggplot representing a dependence plot.
}
\description{
Creates a scatter plot of the SHAP values of a feature against its feature
values. A second variable, color_var, can be selected to be used on the color
axis. In this way, one can get a sense of possible interaction effects. Set
color_var = "auto" to use a simple heuristic to select the color feature with
the strongest apparent interaction. With discrete v, horizontal jitter is
added by default.
}
