% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP.R
\name{shap_force}
\alias{shap_force}
\title{SHAP force plot}
\usage{
shap_force(
  x,
  row.id = 1L,
  max.display = 6L,
  colors = c("#f7d13d", "#a52c60"),
  ...
)
}
\arguments{
\item{x}{An object from \link{shap} function.}

\item{row.id}{A single row number to plot.}

\item{max.display}{Maximum number of features (with largest absolute SHAP
values) should be plotted? If there are more features, they will be collapsed
to one feature. Set to Inf to show all features.}

\item{colors}{A vector of exactly two fill colors: the first for positive
SHAP values, the other for negative ones.}

\item{...}{More arguments.}
}
\value{
An object of class "ggplot" representing a force plot.
}
\description{
Creates a force plot of SHAP values of one single observation. The value of
f(x) denotes the prediction on the SHAP scale, while E(f(x)) refers to the
baseline SHAP value.
}
\examples{
# Examples see in [shap] function.
}
\seealso{
\link{shap_waterfall}
}
