% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\alias{cal.data.frame}
\alias{cal.nmtask}
\alias{cal.glm}
\title{Draw calibrate curve}
\usage{
cal(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  xlab = NULL,
  ylab = NULL,
  show.explain = TRUE,
  ...
)

\method{cal}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  xlab = NULL,
  ylab = NULL,
  show.explain = TRUE,
  ...
)

\method{cal}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  xlab = NULL,
  ylab = NULL,
  show.explain = TRUE,
  ...
)

\method{cal}{glm}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  xlab = NULL,
  ylab = NULL,
  show.explain = TRUE,
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.}

\item{B}{B is an upper limit on the number of resamples for which information
is printed about which variables were selected in each model re-fit, default 1000.}

\item{linesize}{line size, default 0.25.}

\item{linecolor}{line color, the length must be 3.}

\item{linelabel}{line label,the length must be 3.}

\item{xlab}{label for X axis.}

\item{ylab}{label for Y axis.}

\item{show.explain}{explain the figure, default TRUE.}

\item{...}{further arguments.}
}
\description{
Calibration was quantified by comparing the actual observed progression with
the model-predicted progression, and the results were graphically evaluated
as calibration curves. Perfect calibration would be exhibited by a direct
alignment between the actual observation and nomogram prediction probability
along the 45 degree diagonal line.
}
\examples{
head(aps)

# Basic usage
cal(aps,
    outcome = "elope",
    predictors = c("age", "gender", "place3", "neuro"))


# From a nmtask
tk <- nmtask(train.data = aps,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))
cal(tk)


# With validation
index <- sample(1:nrow(aps), 300)
train <- aps[index, ]
test  <- aps[-index, ]

cal(train,
    outcome = "elope",
    predictors = c("age", "gender", "place3", "neuro"),
    newdata = test)
# or
tk <- nmtask(train.data = train,
             test.data = test,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))
cal(tk)
}
