% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Draw calibrate curve}
\usage{
cal(data, outcome, covariates, newdata = NULL)
}
\arguments{
\item{data}{data}

\item{outcome}{outcome}

\item{covariates}{covariates}

\item{newdata}{newdata}
}
\description{
Draw calibrate curve
}
