% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\alias{cal.data.frame}
\alias{cal.nmtask}
\title{Draw calibrate curve}
\usage{
cal(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  ...
)

\method{cal}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  ...
)

\method{cal}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  linesize = 0.5,
  linecolor = NULL,
  linelabel = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.}

\item{B}{B is an upper limit on the number of resamples for which information
is printed about which variables were selected in each model re-fit, default 1000.}

\item{linesize}{line size, default 0.25.}

\item{linecolor}{line color, the length must be 3.}

\item{linelabel}{line label,the length must be 3.}

\item{...}{further arguments.}
}
\description{
Draw calibrate curve
}
