% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefs}
\alias{coefs}
\title{Model coefficients of logistic regression}
\usage{
coefs(
  ...,
  multivariable = TRUE,
  effect.values = c("net", "b", "se", "OR", "p"),
  model.names = NULL,
  conf.level = 0.95,
  conf.brackets = "()",
  conf.separator = NULL,
  digits.pvalue = 3,
  digits.effect = 2,
  ref.value = 1,
  filename = ""
)
}
\arguments{
\item{...}{one or more object of 'nmtask' or 'glm'.}

\item{multivariable}{logical variable indicating whether to perform univariable
logistic analysis or multivariable logistic analysis (default).}

\item{effect.values}{Effect value, 'ne' for No. of event, 'nt' for No. of total,
'net' for No. of event and total, 'nne' for No. of non-event, 'wald' for Wald
Chi-square value, 'B' for coefficients, 'p' for P value, 'se' for standard error,
'or' for odds ratio and its 95\% CI, ignore case.}

\item{model.names}{vector of model's names.}

\item{conf.level}{Effect value level, defalut 0.95.}

\item{conf.brackets}{Confidence interval bracket, default '()'.}

\item{conf.separator}{Confidence interval separator.}

\item{digits.pvalue}{digits for p value, default 3.}

\item{digits.effect}{digits for effect value, default 2.}

\item{ref.value}{Reference Category name, default '1'.}

\item{filename}{filename, if you want to save to word.}
}
\value{
a list.
}
\description{
Model coefficients of logistic regression
}
\examples{
tk <- nmtask(train.data = HCC,
             outcome    = "status",
             predictors = c("AJCC_T", "AJCC_M", "AJCC_N", "Grade", "metastasis",
                            "tumor_size", "Surg_Prim"))

# Multivariable logistic regression, default.
coefs(tk)

# Support for 'glm'
model <- glm(status ~ AJCC_T + AJCC_M + AJCC_N, data = HCC, family = binomial())
coefs(model)

coefs(tk, model, model.names = c("Model 1", "Model 2"))

# Univariable logistic regression
coefs(tk, multivariable = FALSE)
}
