% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca}
\alias{dca}
\alias{dca.data.frame}
\alias{dca.nmtask}
\title{Draw decision curves}
\usage{
dca(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  thresholds = seq(0, 1, by = 0.01),
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{dca}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  thresholds = seq(0, 1, by = 0.01),
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{dca}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  B = 1000,
  thresholds = seq(0, 1, by = 0.01),
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.}

\item{B}{Number of bootstrap replicates to use to calculate confidence intervals (default 1000).}

\item{thresholds}{thresholds.}

\item{xlab}{defaults to "Predicted x-units Survival" or to a suitable label for other models.}

\item{ylab}{defaults to "Fraction Surviving x-units" or to a suitable label for other models.}

\item{...}{further arguments.}
}
\description{
Draw decision curves
}
