% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca}
\alias{dca}
\title{Draw DCA curves}
\usage{
dca(
  ...,
  newdata = NULL,
  model.names = NULL,
  boot = 10,
  thresholds = seq(0, 1, by = 0.01),
  facet = c("data"),
  linewidth = 0.5,
  linecolor = NULL,
  xlab = "Risk threshold",
  ylab = "Standardized net benefit",
  xbreaks = seq(0, 1, 0.2),
  ybreaks = seq(-0.2, 1, 0.2),
  fontfamily = "serif",
  fontsize = 12,
  explain = TRUE,
  seed = 1234
)
}
\arguments{
\item{...}{one or more object of 'nmtask' or 'glm'.}

\item{newdata}{new data for verification.}

\item{model.names}{vector of model's names to use when plotting legends.}

\item{boot}{boot is an upper limit on the number of resamples for which information
is printed about which variables were selected in each model re-fit, default 1000.}

\item{thresholds}{Numeric vector of high risk thresholds to use when plotting
and calculating net benefit values.}

\item{facet}{of 'data' or 'model', specifying grouping variables for faceting
the plot into multiple panels, default 'data'.}

\item{linewidth}{line width, default 0.5.}

\item{linecolor}{line color.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{xbreaks}{x axis breaks.}

\item{ybreaks}{y axis breaks.}

\item{fontfamily}{font family, sefault serif.}

\item{fontsize}{font size, default 12.}

\item{explain}{whether explain the figure, default TRUE.}

\item{seed}{a single value, interpreted as an integer, or NULL, default 1234.}
}
\value{
a ggplot object from the \link{ggplot2} package, or a patchwork object from \link{patchwork} package.
}
\description{
This function calculates decision curves, which are estimates of the standardized
net benefit by the probability threshold used to categorize observations as 'high risk.'
}
\examples{
# Examples see in [nmtask] function.
}
\references{
Vickers AJ, Elkin EB. Decision curve analysis: a novel method for evaluating
prediction models. Med Decis Making. 2006 Nov-Dec;26(6):565-74.
doi: 10.1177/0272989X06295361.
}
