% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtask.R
\name{nmtask}
\alias{nmtask}
\title{Task for nomogram}
\usage{
nmtask(train.data = NULL, test.data = NULL, outcome = NULL, predictors = NULL)
}
\arguments{
\item{train.data}{train data.}

\item{test.data}{test data.}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}
}
\value{
a list.
}
\description{
Task for nomogram
}
