% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtask.R
\name{nmtask}
\alias{nmtask}
\title{Task for nomogram}
\usage{
nmtask(
  train.data,
  test.data = NULL,
  outcome = NULL,
  positive = NULL,
  predictors = NULL
)
}
\arguments{
\item{train.data}{train data.}

\item{test.data}{test data, If there is no test set, it can not be specified.
If specified, it must have the same data structure as train data.}

\item{outcome}{predict outcome.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{predictors}{predictors.}
}
\value{
a object of 'nmtask' class.
}
\description{
Task for nomogram
}
\examples{
index <- sample(1:nrow(aps), 300)
# Training set
train <- aps[index, ]
# Validation set
test  <- aps[-index, ]

# Set task of nomogram model
tk <- nmtask(train.data = train,
             test.data = test,
             outcome = "elope",
             predictors = c("age", "gender", "place3"))
# View task
tk

# Hosmer-Lemeshow test
hl_test(tk)

# Draw a nomogram
nom(tk)

# Draw calibrate curve
cal(tk)

# Draw ROC curve
roc(tk)

# Draw decision curve
# dca(tk)
}
