% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{nom}
\alias{nom}
\alias{nom.data.frame}
\alias{nom.nmtask}
\title{Draw a nomogram}
\usage{
nom(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  ...
)

\method{nom}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  ...
)

\method{nom}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{points.label}{a character string giving the axis label for the points scale.}

\item{total.points.label}{a character string giving the axis label for the total points scale.}

\item{funlabel}{label for fun axis.}

\item{maxscale}{default maximum point score is 100.}

\item{xfrac}{fraction of horizontal plot to set aside for axis titles.}

\item{font.size}{font size.}

\item{fun.at}{function values to label on axis.}

\item{...}{settings of variables to use in constructing axes.}
}
\description{
Draw a nomogram
}
