% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram.R
\name{nom}
\alias{nom}
\title{Draw a nomogram}
\usage{
nom(
  data,
  dependent = NULL,
  independents = NULL,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 11,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{dependent}{dependent variable name.}

\item{independents}{independent variable names.}

\item{points.label}{a character string giving the axis label for the points scale.}

\item{total.points.label}{a character string giving the axis label for the total points scale.}

\item{funlabel}{label for fun axis.}

\item{maxscale}{default maximum point score is 100.}

\item{xfrac}{fraction of horizontal plot to set aside for axis titles.}

\item{font.size}{font size.}

\item{...}{settings of variables to use in constructing axes.}
}
\description{
Draw a nomogram
}
