% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{nom}
\alias{nom}
\title{Draw a nomogram}
\usage{
nom(
  data,
  points.label = "Points",
  total.points.label = "Total points",
  funlabel = "Risk",
  maxscale = 100,
  xfrac = 0.35,
  font.size = 12,
  fun.at = NULL,
  labels = NULL,
  show.points = FALSE,
  show.model = FALSE,
  show.explain = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{points.label}{a character string giving the axis label for the points scale.}

\item{total.points.label}{a character string giving the axis label for the total points scale.}

\item{funlabel}{label for fun axis.}

\item{maxscale}{default maximum point score is 100.}

\item{xfrac}{fraction of horizontal plot to set aside for axis titles.}

\item{font.size}{font size.}

\item{fun.at}{function values to label on axis.}

\item{labels}{a list, labels for variables.}

\item{show.points}{show points, default FALSE..}

\item{show.model}{show model, default FALSE.}

\item{show.explain}{explain the figure, default TRUE.}

\item{...}{settings of variables to use in constructing axes.}
}
\description{
The nomogram can be interpreted as follows: (1) for each  variable,
draw a straight line up to the points axis to determine the points for that
variable, (2) repeat this process for each variable, (3) add the points for
all variables and locate the sum on the total points axis, and (4) draw a
straight line from total points down to risk.
}
\details{
Set labels use \link{set_variable_labels} and \link{set_category_labels} function.
}
\examples{
head(aps)

# From nmtask
tk <- nmtask(train.data = aps,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))

# Set the scale of the risk axis.
nom(tk,
    funlabel = "Risk of Elopement",
    fun.at = c(0.1, 0.15, 0.25, 0.35, 0.45, 0.55, 0.65))

# Fraction of horizontal plot to set aside for axis titles.
nom(tk,
    funlabel = "Risk of Elopement",
    fun.at = c(0.1, 0.15, 0.25, 0.35, 0.45, 0.55, 0.65),
    xfrac = 0.65)

# show model and points
nom(tk,
    funlabel = "Risk of Elopement",
    fun.at = c(0.1, 0.15, 0.25, 0.35, 0.45, 0.55, 0.65),
    xfrac = 0.65,
    show.points = TRUE,
    show.model = TRUE)

nm <- nom(tk,
          funlabel = "Risk of Elopement",
          fun.at = c(0.1, 0.15, 0.25, 0.35, 0.45, 0.55, 0.65),
          xfrac = 0.65)
nm

# set labels
nm |>
  set_variable_labels(age = "Age (years)") |>
  set_variable_labels(gender = "Gender", place3 = "Placement") |>
  set_category_labels("Gender", Male = "M", Female = "F") |>
  set_category_labels("Placement", OutDay = "Outday")
}
