% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Evaluate model performance}
\usage{
performance(..., newdata = NULL, digits = 3, filename = "")
}
\arguments{
\item{...}{one or more object of 'nmtask' or 'glm'.}

\item{newdata}{new data for verification.}

\item{digits}{digits, default 3.}

\item{filename}{filename}
}
\description{
Evaluate model performance
}
