% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{roc.data.frame}
\alias{roc.nmtask}
\title{Draw ROC curve}
\usage{
roc(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  filename = "",
  ...
)

\method{roc}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  filename = "",
  ...
)

\method{roc}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.
is printed about which variables were selected in each model re-fit.}

\item{linesize}{line size, default 0.25.}

\item{linecolor}{line color.}

\item{xlab}{label for X axis.}

\item{ylab}{label for Y axis.}

\item{filename}{file name to save performance metrics.}

\item{...}{further arguments.}
}
\description{
Draw ROC curve
}
