% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{roc.data.frame}
\alias{roc.nmtask}
\alias{roc.glm}
\title{Draw ROC curve}
\usage{
roc(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{roc}{data.frame}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{roc}{nmtask}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{roc}{glm}(
  data,
  outcome = NULL,
  predictors = NULL,
  newdata = NULL,
  linesize = 0.5,
  linecolor = NULL,
  xlab = NULL,
  ylab = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{data}{data}

\item{outcome}{predict outcome.}

\item{predictors}{predictors.}

\item{newdata}{new data for verification.
is printed about which variables were selected in each model re-fit.}

\item{linesize}{line size, default 0.25.}

\item{linecolor}{line color.}

\item{xlab}{label for X axis.}

\item{ylab}{label for Y axis.}

\item{...}{further arguments.}
}
\description{
Draw ROC curve
}
\examples{
head(aps)

# Basic usage
roc(aps,
    outcome = "elope",
    predictors = c("age", "gender", "place3", "neuro"))


# From a nmtask
tk <- nmtask(train.data = aps,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))
roc(tk)


# With validation
index <- sample(1:nrow(aps), 300)
train <- aps[index, ]
test  <- aps[-index, ]

roc(train,
    outcome = "elope",
    predictors = c("age", "gender", "place3", "neuro"),
    newdata = test)
# or
tk <- nmtask(train.data = train,
             test.data = test,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))
roc(tk)
}
