% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{set_category_labels}
\alias{set_category_labels}
\title{Set labels for category of variable}
\usage{
set_category_labels(x, variable, ...)
}
\arguments{
\item{x}{a nomogram object from \link{nom} function}

\item{variable}{variable name from nomogram.}

\item{...}{"category=label"}
}
\value{
a nomogram.
}
\description{
Set labels for category of variable
}
\examples{
nm <- nom(data = aps,
          outcome = "elope",
          predictors = c("age", "gender", "place3"))

nm |>
  set_variable_labels(age = "Age (years)") |>
  set_variable_labels(gender = "Gender", place3 = "Placement") |>
  set_category_labels("Gender", Male = "M", Female = "F") |>
  set_category_labels("Placement", OutDay = "Outday")
}
\seealso{
\link{set_variable_labels}
}
