% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{set_variable_labels}
\alias{set_variable_labels}
\title{Set labels for nomogram}
\usage{
set_variable_labels(x, ...)
}
\arguments{
\item{x}{a nomogram object from nom() function}

\item{...}{"vairable=label"}
}
\value{
a nomogram.
}
\description{
Set labels for nomogram
}
\examples{
tk <- nmtask(train.data = aps,
             outcome = "elope",
             predictors = c("age", "gender", "place3", "neuro"))

nm <- nom(tk,
          funlabel = "Risk of Elopement",
          fun.at = c(0.1, 0.15, 0.25, 0.35, 0.45, 0.55, 0.65),
          xfrac = 0.65)
nm

# set labels
nm |>
  set_variable_labels(age = "Age (years)") |>
  set_variable_labels(gender = "Gender", place3 = "Placement") |>
  set_category_labels("Gender", Male = "M", Female = "F") |>
  set_category_labels("Placement", OutDay = "Outday")
}
\seealso{
\link{set_category_labels}
}
