% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP.R
\name{shap_force}
\alias{shap_force}
\title{SHAP force plot}
\usage{
shap_force(x, row_id = 1L, colors = c("#f7d13d", "#a52c60"), ...)
}
\arguments{
\item{x}{An object from \link{shap} function.}

\item{row_id}{A single row number to plot.}

\item{colors}{A vector of exactly two fill colors: the first for positive
SHAP values, the other for negative ones.}

\item{...}{More arguments.}
}
\value{
An object of class "ggplot" representing a force plot.
}
\description{
Creates a force plot of SHAP values of one single observation. The value of
f(x) denotes the prediction on the SHAP scale, while E(f(x)) refers to the
baseline SHAP value.
}
\seealso{
\link{shap_waterfall}
}
